Switch to diffing approach
