+++ app/reducers/gridReducer.ts
@@ -40,7 +40,6 @@ import {
   postEdit as builderPostEdit,
   clearSelection,
   hasSelection,
-  pushToHistory,
 } from './builderUtils';
 import type { PuzzleState } from './puzzleReducer';
 import { isPuzzleState, postEdit as puzzlePostEdit } from './puzzleUtils';
@@ -244,16 +243,12 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
           state = enterCharAt(state, pos, EMPTY, Symmetry.None);
         }
       });
-      if (isCutAction(action)) {
-        state = pushToHistory(state);
-      }
     } else {
       const val = valAt(state.grid, state.active);
       if (val !== BLOCK && val !== EMPTY) {
         toCopy = val;
         if (isCutAction(action)) {
           state = enterText(state, EMPTY);
-          state = pushToHistory(state);
         }
       }
     }
@@ -295,7 +290,6 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
         }
       });
     });
-    state = pushToHistory(state);
     state = {
       ...state,
       wasEntryClick: false,
@@ -336,7 +330,6 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
         } else {
           state.grid.highlighted.add(ci);
         }
-        state = pushToHistory(state);
       }
       return { ...state };
     }
@@ -349,7 +342,7 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
           rebusValue: state.rebusValue ? state.rebusValue.slice(0, -1) : '',
         };
       } else if (key.k === KeyK.Enter) {
-        state = {
+        return {
           ...closeRebus(state),
           wasEntryClick: false,
           active: advancePosition(
@@ -359,8 +352,6 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
             isPuzzleState(state) ? state.prefs : undefined
           ),
         };
-        state = pushToHistory(state);
-        return state;
       } else if (key.k === KeyK.Escape) {
         return { ...state, isEnteringRebus: false, rebusValue: '' };
       }
@@ -526,13 +517,11 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
         const symmetry = isBuilderState(state) ? state.symmetry : Symmetry.None;
         const grid = gridWithBlockToggled(state.grid, state.active, symmetry);
         state = clearSelection(state);
-        state = {
+        return {
           ...postEdit({ ...state, grid }, ci),
           wasEntryClick: false,
           active: nextCell(state.grid, state.active),
         };
-        state = pushToHistory(state);
-        return state;
       }
       return state;
     } else if (key.k === KeyK.Comma && state.grid.allowBlockEditing) {
@@ -541,18 +530,15 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
         const symmetry = isBuilderState(state) ? state.symmetry : Symmetry.None;
         const grid = gridWithBarToggled(state.grid, state.active, symmetry);
         state = clearSelection(state);
-        state = {
+        return {
           ...postEdit({ ...state, grid }, ci),
           wasEntryClick: false,
         };
-        state = pushToHistory(state);
-        return state;
       }
       return state;
     } else if (key.k === KeyK.AllowedCharacter) {
       const char = key.c.toUpperCase();
       state = enterText(state, char);
-      state = pushToHistory(state);
       return {
         ...state,
         wasEntryClick: false,
@@ -583,7 +569,6 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
         }
       }
       state = clearSelection(state);
-      state = pushToHistory(state);
       return {
         ...state,
         wasEntryClick: false,
@@ -609,7 +594,6 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
         }
       }
       state = clearSelection(state);
-      state = pushToHistory(state);
       return {
         ...state,
         wasEntryClick: false,
@@ -620,14 +604,11 @@ export function gridInterfaceReducer<T extends GridInterfaceState>(
       if (state.isEditable(ci)) {
         const symmetry = isBuilderState(state) ? state.symmetry : Symmetry.None;
         const grid = gridWithHiddenToggled(state.grid, state.active, symmetry);
-        state = {
+        return {
           ...postEdit({ ...state, grid }, ci),
           wasEntryClick: false,
         };
-        state = pushToHistory(state);
-        return state;
       }
-      return state;
     }
   }
   return state;
